refactor: migrate errors implementation into drift-sdk
(cherry picked from commit 15ebdf0aa087065de1f5b57181d45cc4cc5766ef)
